refactor : 사용하지 않는 이전 코드 삭제
