--- server/collusic-be/src/main/java/com/collusic/collusicbe/domain/contributeproject/ContributeProjectRepository.java
@@ -1,13 +1,0 @@
-package com.collusic.collusicbe.domain.contributeproject;
-
-import org.springframework.data.domain.Sort;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface ContributeProjectRepository extends JpaRepository<ContributeProject, Long> {
-
-    List<ContributeProject> findByRequestProjectId(Long requestProjectId);
-
-    List<ContributeProject> findAllByRequestProjectId(Long requestProjectId, Sort sort);
-}
