--- server/collusic-be/src/main/java/com/collusic/collusicbe/domain/field/RequestProjectFieldEntity.java
@@ -1,23 +1,0 @@
-package com.collusic.collusicbe.domain.field;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@NoArgsConstructor
-@Table(name = "REQUEST_PROJECT_FIELD")
-@Getter
-public class RequestProjectFieldEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String field;
-
-    public RequestProjectFieldEntity(String field) {
-        this.field = field;
-    }
-}
