--- server/collusic-be/src/main/java/com/collusic/collusicbe/domain/genre/GenreEntity.java
@@ -1,25 +1,0 @@
-package com.collusic.collusicbe.domain.genre;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@NoArgsConstructor
-@Getter
-@Table(name = "GENRE")
-public class GenreEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String genre;
-
-    public GenreEntity(String genre) {
-        this.genre = genre;
-    }
-}
-
-
