--- server/collusic-be/src/main/java/com/collusic/collusicbe/domain/mood/MoodEntity.java
@@ -1,23 +1,0 @@
-package com.collusic.collusicbe.domain.mood;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@NoArgsConstructor
-@Getter
-@Table(name = "MOOD")
-public class MoodEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String mood;
-
-    public MoodEntity(String mood) {
-        this.mood = mood;
-    }
-}
