--- server/collusic-be/src/main/java/com/collusic/collusicbe/domain/requestproject/RequestProjectRepository.java
@@ -1,6 +1,0 @@
-package com.collusic.collusicbe.domain.requestproject;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RequestProjectRepository extends JpaRepository<RequestProject, Long> {
-}
