--- server/collusic-be/src/main/java/com/collusic/collusicbe/global/exception/jwt/TokenNotFoundException.java
@@ -1,10 +1,0 @@
-package com.collusic.collusicbe.global.exception.jwt;
-
-import io.jsonwebtoken.JwtException;
-
-public class TokenNotFoundException extends JwtException {
-
-    public TokenNotFoundException(String message) {
-        super(message);
-    }
-}
