+++ server/collusic-be/src/main/java/com/collusic/collusicbe/service/S3Service.java
@@ -50,27 +50,6 @@ public class S3Service {
         return cloudFrontDomain + path;
     }
 
-    public String upload(@ModelAttribute MultipartFile multipartFile, String dirName) throws IOException {
-        File uploadFile = convert(multipartFile)
-                .orElseThrow(() -> new IllegalArgumentException("MultipartFile -> File로 전환이 실패했습니다."));
-
-        return upload(uploadFile, dirName);
-    }
-
-    private String upload(File uploadFile, String dirName) {
-        String fileName = dirName + "/" + uploadFile.getName();
-        String uploadImageUrl = putS3(uploadFile, fileName);
-        removeNewFile(uploadFile);
-        return uploadImageUrl;
-    }
-
-    public String update(@ModelAttribute MultipartFile multipartFile, String dirName, String savedFileName) throws IOException {
-        File uploadFile = convert(multipartFile)
-                .orElseThrow(() -> new IllegalArgumentException("MultipartFile -> File로 전환이 실패했습니다."));
-
-        return update(uploadFile, dirName, savedFileName);
-    }
-
     private String update(File uploadFile, String dirName, String savedFileName) {
         String fileName = dirName + "/" + uploadFile.getName();
         if (isExist(savedFileName)) {
