--- server/collusic-be/src/main/java/com/collusic/collusicbe/util/StringUtils.java
@@ -1,8 +1,0 @@
-package com.collusic.collusicbe.util;
-
-public class StringUtils {
-    public static String extractFileNameFromFilePath(String filePath) {
-        int startIndex = filePath.lastIndexOf("static");
-        return filePath.substring(startIndex);
-    }
-}
