+++ server/collusic-be/src/main/java/com/collusic/collusicbe/web/controller/ProjectsResponseDto.java
@@ -1,5 +1,6 @@
 package com.collusic.collusicbe.web.controller;
 
+import com.collusic.collusicbe.web.controller.dto.ProjectPreview;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
