--- server/collusic-be/src/main/java/com/collusic/collusicbe/web/dto/RequestProjectsWithPaginationDto.java
@@ -1,18 +1,0 @@
-package com.collusic.collusicbe.web.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Getter
-@NoArgsConstructor
-public class RequestProjectsWithPaginationDto {
-    private List<RequestProjectResponseDto> requestProjectResponseDtos;
-    private int totalPages;
-
-    public RequestProjectsWithPaginationDto(List<RequestProjectResponseDto> requestProjectResponseDtos, int totalPages) {
-        this.requestProjectResponseDtos = requestProjectResponseDtos;
-        this.totalPages = totalPages;
-    }
-}
